Move ResizeStatus to it's own file
